moodwave-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.1.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
firebase-admin==6.5.0
httpx==0.27.2
python-dotenv==1.0.1
spotipy==2.24.0
numpy==1.26.4
slowapi==0.1.9
APScheduler==3.10.4
fastapi-mail==1.4.1
